+++ src/main/java/blackjack/BlackJackGameBoard.java
@@ -1,4 +1,4 @@
-package blackjack.controller;
+package blackjack;
 
 import blackjack.domain.BetManager;
 import blackjack.domain.Game;
